old-tests/flaky.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Flaky test</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  assert_true(Math.random() > 0.5, "Flaky failure");
});
</script>
